% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataVar.R
\name{getDataVar}
\alias{getDataVar}
\title{Get Data from WS2 and formate them}
\usage{
getDataVar(varURI, variableList, token)
}
\arguments{
\item{varURI}{uri of the variable to plot, from the \code{\link{variableList}} function or the web service directly}

\item{variableList}{from \code{\link{variableList}}}

\item{token}{a token from \code{\link{getToken}} function}
}
\value{
WSResponse
}
\description{
Get Data from WS2 and formate them
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 varPrettyTot <- variableList(token = token)
 getDataVar(varURI = listVariables(token,
                  wsUrl="www.opensilex.org/openSilexAPI/rest/")$value[1],
                  variableList = varPrettyTot,
                  token = token)
}
}
